Fixes after upgrade engine from Abrodot to Godot

+++ Scenes/Screen/Components/TwoColoredBar/TwoColoredBar.cs
@@ -10,13 +10,13 @@ public partial class TwoColoredBar : ColorRect
 	[Export] [NotNull] public ColorRect LowerBar { get; private set; }
 	[Export] [NotNull] public Label Label { get; private set; }
 	
-	public double MaxValue { get; set; }
-	public double CurrentUpperValue { get; set; }
-	public double CurrentLowerValue { get; set; }
+	public float MaxValue { get; set; }
+	public float CurrentUpperValue { get; set; }
+	public float CurrentLowerValue { get; set; }
 	
-	public double CurrentUpperValuePercent => CurrentUpperValue / MaxValue;
-	public double CurrentLowerValuePercent => CurrentLowerValue / MaxValue;
-	public double Width => Size.X;
+	public float CurrentUpperValuePercent => CurrentUpperValue / MaxValue;
+	public float CurrentLowerValuePercent => CurrentLowerValue / MaxValue;
+	public float Width => Size.X;
 	
 	public override void _Ready()
 	{
